unit test mock cleanup and cleanup default router
